requirements.txt
django==3.8
django-bootstrap4==23.1
django-cors-headers==3.11.0
django-forms-bootstrap==3.1.0
djangorestframework==3.12
gunicorn==20.1
jinja2==2.1.1
psycopg2-binary==2.9.3
pytz==2021.3
whitenoise==5.3.0
pillow==9.4.0

